Backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-filter
djangorestframework
djangorestframework_simplejwt
djoser
drf-nested-routers
Faker
idna
oauthlib
pillow
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
tzdata
urllib3
gunicorn
